(long long) and %lld for big time_t ok nicm, who will pass it upstream

+++ lib/libevent/event.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: event.c,v 1.26 2012/08/28 09:09:56 pascal Exp $	*/
+/*	$OpenBSD: event.c,v 1.27 2013/04/17 15:31:49 deraadt Exp $	*/
 
 /*
  * Copyright (c) 2000-2004 Niels Provos <provos@citi.umich.edu>
@@ -794,8 +794,8 @@ event_add(struct event *ev, const struct timeval *tv)
 		evutil_timeradd(&now, tv, &ev->ev_timeout);
 
 		event_debug((
-			 "event_add: timeout in %ld seconds, call %p",
-			 tv->tv_sec, ev->ev_callback));
+			 "event_add: timeout in %lld seconds, call %p",
+			 (long long)tv->tv_sec, ev->ev_callback));
 
 		event_queue_insert(base, ev, EVLIST_TIMEOUT);
 	}
@@ -884,7 +884,7 @@ timeout_next(struct event_base *base, struct timeval **tv_p)
 	assert(tv->tv_sec >= 0);
 	assert(tv->tv_usec >= 0);
 
-	event_debug(("timeout_next: in %ld seconds", tv->tv_sec));
+	event_debug(("timeout_next: in %lld seconds", (long long)tv->tv_sec));
 	return (0);
 }
